+++ R/model.R
@@ -702,8 +702,8 @@ evaluate_predictor <- function(model,
 #' @return A vector of values for a component
 #' @examples
 #' if (bru_safe_inla() &&
-#'     require("sf", quietly = TRUE) &&
-#'     requireNamespace("sn", quietly = TRUE)) {
+#'   require("sf", quietly = TRUE) &&
+#'   requireNamespace("sn", quietly = TRUE)) {
 #'   mesh <- fmesher::fm_mesh_2d_inla(
 #'     cbind(0, 0),
 #'     offset = 2,